error handling

+++ src/main/java/com/example/FizzBuzz/service/SequenceService.java
@@ -2,8 +2,10 @@ package com.example.FizzBuzz.service;
 
 import com.example.FizzBuzz.controller.response.SequenceResponse;
 
+import java.util.List;
+
 public interface SequenceService {
 
-    SequenceResponse generateSequence(long lastElement);
+    SequenceResponse generateSequence(int lastElement);
 
 }
